Add sway-focus-next
sway-focus-next

It focuses the next in the line based on the focused windows (or the
first matching the criteria). It was based on sway-focus-next-visible,
but I've made it more general than that, which means that you must
specify visible as a criteria.

It uses a crude but effective expression language for the criteria
which amounts to `$path$op$rval`.

The $path is delimited by `/` (almost like folders) and currently only
considers objects. The spaces are significant and it has no mechanism
for escaping `/`. Trailing spaces are significant as well, and so
it must be immediately followed by an operator.

Operators include `>=`, `<=`, `<`, `>` (only on numbers), `!=`, `^=`
(startswith), `$=` (endswith).

The target value is a json value.

Multiple clauses are `&&`'d together.

Examples:

`sway-focus-next visible==true 'window_properties/class^="st"'`
`sway-focus-next visible==true 'type=='(sway-focused-window type)`
`sway-focus-next 'type=="con"'`

Add utils module for common code.